+++ utils/cache_manager.py
@@ -1,0 +1,7 @@
+from flask_caching import Cache
+from utils.cache_config import cache_config
+
+cache = Cache()
+
+def init_cache(app):
+    cache.init_app(app, config=cache_config)
